Refactor MapTile into it's own file
